fully support float zoom value. Right now when in between int value the tile is scaled and so not as sharp in the lowest zoom (to see the full map). Not sure how to solve this ...

+++ MapboxAndroidSDK/src/main/java/com/mapbox/mapboxsdk/tileprovider/modules/MapTileModuleLayerBase.java
@@ -61,14 +61,14 @@ public abstract class MapTileModuleLayerBase implements TileLayerConstants {
      *
      * @return the minimum zoom level
      */
-    public abstract int getMinimumZoomLevel();
+    public abstract float getMinimumZoomLevel();
 
     /**
      * Gets the maximum zoom level this tile provider can provide
      *
      * @return the maximum zoom level
      */
-    public abstract int getMaximumZoomLevel();
+    public abstract float getMaximumZoomLevel();
 
     /**
      * Sets the tile source for this tile provider.
